Matt's Organic Chemistry I Curriculum
This curriculum comes from two text books.  One is a cellular biology book and the other is an organic chemistry textbook.  Each quiz comes straight out of the organic chemistry workbook where marks are delegated accordingly to the answers.  There are 2 midterms, the course component (chapter quizzes) and a final exam.